refactor: Move model serving datas to data package
