+++ src/ai/backend/manager/data/model_serving/modifier.py
@@ -1,8 +1,11 @@
 import uuid
 from dataclasses import dataclass, field
+from decimal import Decimal
 from typing import Any, override
 
 from ai.backend.common.types import (
+    AutoScalingMetricComparator,
+    AutoScalingMetricSource,
     ClusterMode,
     ResourceSlot,
     RuntimeVariant,
@@ -72,3 +75,30 @@ class EndpointModifier(PartialModifier):
 
     def replica_count_modified(self) -> bool:
         return self.replicas.optional_value() is not None
+
+
+@dataclass
+class EndpointAutoScalingRuleModifier(PartialModifier):
+    metric_source: OptionalState[AutoScalingMetricSource] = field(default_factory=OptionalState.nop)
+    metric_name: OptionalState[str] = field(default_factory=OptionalState.nop)
+    threshold: OptionalState[Decimal] = field(default_factory=OptionalState.nop)
+    comparator: OptionalState[AutoScalingMetricComparator] = field(
+        default_factory=OptionalState.nop
+    )
+    step_size: OptionalState[int] = field(default_factory=OptionalState.nop)
+    cooldown_seconds: OptionalState[int] = field(default_factory=OptionalState.nop)
+    min_replicas: TriState[int] = field(default_factory=TriState.nop)
+    max_replicas: TriState[int] = field(default_factory=TriState.nop)
+
+    @override
+    def fields_to_update(self) -> dict[str, Any]:
+        to_update: dict[str, Any] = {}
+        self.metric_source.update_dict(to_update, "metric_source")
+        self.metric_name.update_dict(to_update, "metric_name")
+        self.threshold.update_dict(to_update, "threshold")
+        self.comparator.update_dict(to_update, "comparator")
+        self.step_size.update_dict(to_update, "step_size")
+        self.cooldown_seconds.update_dict(to_update, "cooldown_seconds")
+        self.min_replicas.update_dict(to_update, "min_replicas")
+        self.max_replicas.update_dict(to_update, "max_replicas")
+        return to_update
