+++ src/ai/backend/manager/repositories/deployment/storage_source/storage_source.py
@@ -5,10 +5,10 @@ from typing import Optional
 import tomli
 
 from ai.backend.common.types import VFolderID
+from ai.backend.manager.data.model_serving.types import ModelServiceDefinition
 from ai.backend.manager.data.vfolder.types import VFolderLocation
 from ai.backend.manager.errors.deployment import DefinitionFileNotFound
 from ai.backend.manager.models.storage import StorageSessionManager
-from ai.backend.manager.services.model_serving.types import ModelServiceDefinition
 
 
 class DeploymentStorageSource:
