+++ src/ai/backend/manager/services/model_serving/actions/create_model_service.py
@@ -3,8 +3,9 @@ from dataclasses import dataclass
 from typing import Optional, override
 
 from ai.backend.manager.actions.action import BaseActionResult
+from ai.backend.manager.data.model_serving.creator import ModelServiceCreator
+from ai.backend.manager.data.model_serving.types import ServiceInfo
 from ai.backend.manager.services.model_serving.actions.base import ModelServiceAction
-from ai.backend.manager.services.model_serving.types import ModelServiceCreator, ServiceInfo
 
 
 @dataclass
