+++ src/ai/backend/manager/services/model_serving/actions/dry_run_model_service.py
@@ -6,8 +6,8 @@ from pydantic import AnyUrl
 
 from ai.backend.common.types import ClusterMode, RuntimeVariant
 from ai.backend.manager.actions.action import BaseActionResult
+from ai.backend.manager.data.model_serving.types import ModelServicePrepareCtx, ServiceConfig
 from ai.backend.manager.services.model_serving.actions.base import ModelServiceAction
-from ai.backend.manager.services.model_serving.types import ModelServicePrepareCtx, ServiceConfig
 
 
 @dataclass
