+++ tests/manager/services/model_serving/actions/test_update_route.py
@@ -4,6 +4,7 @@ from unittest.mock import AsyncMock, MagicMock
 import aiohttp
 import pytest
 
+from ai.backend.manager.data.model_serving.types import RequesterCtx
 from ai.backend.manager.errors.service import ModelServiceNotFound
 from ai.backend.manager.models.user import UserRole
 from ai.backend.manager.services.model_serving.actions.update_route import (
@@ -13,7 +14,6 @@ from ai.backend.manager.services.model_serving.actions.update_route import (
 from ai.backend.manager.services.model_serving.processors.model_serving import (
     ModelServingProcessors,
 )
-from ai.backend.manager.services.model_serving.types import RequesterCtx
 
 from ...utils import ScenarioBase
